+++ htagweb/session/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# #############################################################################
+# Copyright (C) 2023 manatlan manatlan[at]gmail(dot)com
+#
+# MIT licence
+#
+# https://github.com/manatlan/htagweb
+# #############################################################################
+
+from .sessionfile import create as createFile
+from .shm import create as createShm
+
+__all__= ["createFile","createShm"]
