+++ htagweb/session/shm.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# #############################################################################
+# Copyright (C) 2023 manatlan manatlan[at]gmail(dot)com
+#
+# MIT licence
+#
+# https://github.com/manatlan/htagweb
+# #############################################################################
+
+def create(uid,size=10240):
+    from shared_memory_dict import SharedMemoryDict
+    return SharedMemoryDict(name=uid, size=10240)
